Refactor sidebar components and add tests
